web_app/requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4
celery[redis]
Django==3.1
importlib-metadata==1.7.0
kombu==4.6.11
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
gunicorn
gevent
names
py-loremipsum
bootstrap-admin
